Feature/single folder build

+++ src/engine/server/routes/routes.py
@@ -9,6 +9,7 @@ from flask_cors import cross_origin
 # pylint: disable=E0401, E0611
 # pylint: disable=cyclic-import
 # pylint: disable=undefined-variable
+
 from server import app, global_store, server, wait_for_process
 from server.services import decode, encode
 
@@ -92,7 +93,8 @@ def shutdown():
     """
     Shutdown the server
     """
-    global_store.delete_cache()
+    if request.args.get("deleteCache"):
+        global_store.delete_cache()
     server.socket.stop()
     return "Server shutting down...", 200
 
@@ -136,6 +138,34 @@ def reject_cache() -> Response:
     """Set cache is called."""
     global_store.cache_is_called()
     return Response(status=200)
+
+
+@app.route("/set_gpu", methods=["POST"])
+@cross_origin()
+@portal_function_handler(clear_status=False)
+def set_gpu() -> Response:
+    """Set the GPU flag to true."""
+    with open(os.getenv("GPU_DIR"), "w") as gpu_flag:
+        gpu_flag.write("0")
+    return Response(status=200)
+
+
+@app.route("/clear_gpu", methods=["POST"])
+@cross_origin()
+@portal_function_handler(clear_status=False)
+def clear_gpu() -> Response:
+    """Clear the GPU flag."""
+    with open(os.getenv("GPU_DIR"), "w") as gpu_flag:
+        gpu_flag.write("-1")
+    return Response(status=200)
+
+
+@app.route("/get_gpu", methods=["GET"])
+@cross_origin()
+@portal_function_handler(clear_status=False)
+def get_gpu() -> Response:
+    """Get the GPU flag."""
+    return Response(status=200, response=os.getenv("CUDA_VISIBLE_DEVICES"))
 
 
 @app.route("/api/model/register", methods=["POST"])
